CLE_007: Change to Clean Architecture
