+++ .env.development
@@ -1,0 +1,4 @@
+AUTH_SECRET="8CyBOF6iZrYSs1CKDSS1Nlz2RgyuGPhQs7dYhKK83G4="
+MONGODB_DB="Charles_learn"
+MONGODB_URI="mongodb+srv://quanminh:minhquan@cluster0.x0aqi9k.mongodb.net/"
+# MONGODB_URI="mongodb://localhost:27017"
