+++ .vscode/launch.json
@@ -15,18 +15,14 @@
     },
     {
       "name": "Next.js: debug full stack",
-      "type": "node",
+      "type": "node-terminal",
       "request": "launch",
-      "program": "${workspaceFolder}/node_modules/.bin/next",
-      "runtimeArgs": ["--inspect"],
-      "skipFiles": ["<node_internals>/**"],
+      "command": "${workspaceFolder}/node_modules/next/dist/bin/next",
       "serverReadyAction": {
-        "action": "debugWithEdge",
-        "killOnServerStop": true,
         "pattern": "- Local:.+(https?://.+)",
         "uriFormat": "%s",
-        "webRoot": "${workspaceFolder}"
+        "action": "debugWithChrome"
       }
-    }
+    },
   ]
 }
