+++ Dockerfile
@@ -18,4 +18,4 @@ RUN npm install
 EXPOSE 3000
 
 # Start the app using serve command
-CMD [ "npm", "run", "dev" ]
+CMD [ "npm", "run", "start" ]
