+++ app/(authentication)/list-word/add-word.tsx
@@ -2,7 +2,7 @@ import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { Form, FormControl, FormField, FormItem, FormLabel } from '@/components/ui/form'
 import { Input } from '@/components/ui/input'
-import { useWordStore } from '@/stores/useListWord'
+import { useWordStore } from '$root/stores/useListWord'
 import React from 'react'
 import { useForm } from 'react-hook-form'
 import { z } from "zod"
