+++ auth.config.ts
@@ -1,4 +1,5 @@
 import type { NextAuthConfig } from 'next-auth';
+import { NextResponse } from 'next/server';
 
 export const authConfig = {
     pages: {
@@ -9,20 +10,20 @@ export const authConfig = {
             const isLoggedIn = !!auth?.user;
             const isOnLoginPage = nextUrl.pathname === '/login';
             const isOnLanding = nextUrl.pathname === '/';
-            
-            if (isLoggedIn && isOnLoginPage && !isOnLanding) {
-                // Nếu đã đăng nhập và đang ở trang login, chuyển hướng tới trang chủ
-                return Response.redirect(new URL('/home', nextUrl));
+            const isAuthPage = isOnLoginPage || isOnLanding;
+
+            // Đã đăng nhập -> chuyển về home nếu đang ở trang auth
+            if (isLoggedIn && isAuthPage) {
+                return NextResponse.redirect(new URL('/home', nextUrl));
             }
 
-            if (!isLoggedIn && !isOnLoginPage && !isOnLanding) {
-                // Nếu chưa đăng nhập và không ở trang login, chuyển hướng tới trang login
-                return Response.redirect(new URL('/login', nextUrl));
+            // Chưa đăng nhập -> chuyển về login nếu không ở trang auth
+            if (!isLoggedIn && !isAuthPage) {
+                return NextResponse.redirect(new URL('/login', nextUrl));
             }
 
-            // Mặc định, cho phép truy cập trang
             return true;
         },
     },
-    providers: [],
+    providers: [], // Providers sẽ được định nghĩa trong auth.ts
 } satisfies NextAuthConfig;
