--- lib/data-list-word.js
@@ -1,102 +1,0 @@
-const listWord = [
-  {
-    "id": 1,
-    "english_word": "apple",
-    "vietnamese_word": "táo"
-  },
-  {
-    "id": 2,
-    "english_word": "banana",
-    "vietnamese_word": "chuối"
-  },
-  {
-    "id": 3,
-    "english_word": "cat",
-    "vietnamese_word": "mèo"
-  },
-  {
-    "id": 4,
-    "english_word": "dog",
-    "vietnamese_word": "chó"
-  },
-  {
-    "id": 5,
-    "english_word": "elephant",
-    "vietnamese_word": "voi"
-  },
-  {
-    "id": 6,
-    "english_word": "fish",
-    "vietnamese_word": "cá"
-  },
-  {
-    "id": 7,
-    "english_word": "grape",
-    "vietnamese_word": "nho"
-  },
-  {
-    "id": 8,
-    "english_word": "house",
-    "vietnamese_word": "nhà"
-  },
-  {
-    "id": 9,
-    "english_word": "ice",
-    "vietnamese_word": "đá"
-  },
-  {
-    "id": 10,
-    "english_word": "juice",
-    "vietnamese_word": "nước ép"
-  },
-  {
-    "id": 11,
-    "english_word": "kite",
-    "vietnamese_word": "diều"
-  },
-  {
-    "id": 12,
-    "english_word": "lion",
-    "vietnamese_word": "sư tử"
-  },
-  {
-    "id": 13,
-    "english_word": "monkey",
-    "vietnamese_word": "khỉ"
-  },
-  {
-    "id": 14,
-    "english_word": "notebook",
-    "vietnamese_word": "sổ tay"
-  },
-  {
-    "id": 15,
-    "english_word": "orange",
-    "vietnamese_word": "cam"
-  },
-  {
-    "id": 16,
-    "english_word": "pencil",
-    "vietnamese_word": "bút chì"
-  },
-  {
-    "id": 17,
-    "english_word": "queen",
-    "vietnamese_word": "nữ hoàng"
-  },
-  {
-    "id": 18,
-    "english_word": "rabbit",
-    "vietnamese_word": "thỏ"
-  },
-  {
-    "id": 19,
-    "english_word": "snake",
-    "vietnamese_word": "rắn"
-  },
-  {
-    "id": 20,
-    "english_word": "tiger",
-    "vietnamese_word": "hổ"
-  }
-]
