+++ types/auth.d.ts
@@ -1,0 +1,18 @@
+import { DefaultSession } from 'next-auth';
+
+declare module 'next-auth' {
+    interface Session {
+        user: {
+            id: string;
+            email: string;
+            name: string;
+        } & DefaultSession['user']
+    }
+
+    interface User {
+        id: string;
+        email: string;
+        name: string;
+        password: string;
+    }
+}
